code_file/requirements.txt
datasets==3.1.0
import_ipynb==0.2
llama_index==0.12.1
pandas==2.2.3
protobuf==3.20.3
python-dotenv==1.0.1
ragas==0.2.6
tiktoken==0.8.0